requirements.txt
dearpygui
requests
pillow
numpy
